Move validation to validators.py.
