+++ backend/api/serializers.py
@@ -1,5 +1,4 @@
 from django.db import transaction
-from django.utils import timezone
 from djoser.serializers import UserCreateSerializer, UserSerializer
 from rest_framework import serializers
 from taggit.models import Tag
@@ -24,7 +23,11 @@ from projects.models import (
 )
 from users.models import User
 
-from .validators import validate_status_incomes
+from .validators import (
+    validate_dates,
+    validate_reception_status,
+    validate_status_incomes,
+)
 
 
 class AddressSerializer(serializers.ModelSerializer):
@@ -154,62 +157,14 @@ class ProjectSerializer(serializers.ModelSerializer):
     # category = ProjectCategorySerializer()
     # city = CitySerializer()
 
-    def validate_dates(self, start, end, application):
-        """
-        Проверяет корректность дат.
-        """
-        now = timezone.now()
-
-        if start <= now:
-            raise serializers.ValidationError(
-                'Дата начала мероприятия должна быть в будущем.'
-            )
-        if end <= start:
-            raise serializers.ValidationError(
-                'Дата окончания мероприятия должна быть позже даты начала.'
-            )
-        if not (application <= start <= end):
-            raise serializers.ValidationError(
-                'Дата подачи заявки должна быть позже или равна дате начала '
-                'мероприятия и позже даты начала и раньше даты окончания.'
-            )
-        return start, end, application
-
-    def validate_reception_status(
-        self, status_project, application_date, start_datetime, end_datetime
-    ):
-        """
-        Проверяет, что статус "Прием откликов окончен" можно устанавливать
-        только после указанной даты подачи заявки.
-        """
-        now = timezone.now()
-        if status_project == Project.RECEPTION_OF_RESPONSES_CLOSED:
-            if application_date > now:
-                raise serializers.ValidationError(
-                    'Статус проекта "Прием откликов окончен" можно установить'
-                    'только после окончания подачи заявок.'
-                )
-        if status_project == Project.READY_FOR_FEEDBACK:
-            if now < start_datetime or now < application_date:
-                raise serializers.ValidationError(
-                    'Статус проекта "Готов к откликам" можно установить до '
-                    'начала мероприятия и до даты подачи заявки.'
-                )
-        if status_project == Project.PROJECT_COMPLETED:
-            if now < end_datetime:
-                raise serializers.ValidationError(
-                    'Статус проекта "Проект завершен" можно установить '
-                    'только после окончания мероприятия.'
-                )
-
     def validate(self, data):
         start_datetime = data['start_datetime']
         end_datetime = data['end_datetime']
         application_date = data['application_date']
         status_project = data.get('status_project')
 
-        self.validate_dates(start_datetime, end_datetime, application_date)
-        self.validate_reception_status(
+        validate_dates(start_datetime, end_datetime, application_date)
+        validate_reception_status(
             status_project, application_date, start_datetime, end_datetime
         )
         return data
